+++ src/main/java/me/drex/instantfeedback/duck/snow_golem/ISnowGolem.java
@@ -1,0 +1,6 @@
+package me.drex.instantfeedback.duck.snow_golem;
+
+public interface ISnowGolem {
+    boolean instantfeedback$hasPalePumpkin();
+    void instantfeedback$setPalePumpkin(boolean palePumpkin);
+}